backend/meeting_summary_tool/backend/requirements.txt
fastapi
uvicorn
python-multipart
whisper
python-dotenv
langchain
langchain-google-genai